requirements.txt
selenium==4.8.0
webdriver-manager==4.0.2
requests==2.31.0
pyttsx3==2.91
pygame~=2.6.0
python-dotenv~=1.0.1